Add initial a11y site
This site contains dummy blog posts, and chatGPT content for the
entirety of the site. The purpose of the site is to get it in front of
G and B for a quick go over. We will then focus on modifying the service
offerings, fixxing accessibility errors, and giving a list of blog
topics to work on.

inital availability at frenetic.workyardlabs.com